+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQA457T59412PHRADRWSXTRC"
+        "01KQAEKN39YWV55M3EE1FM8SHN"
     )
     r"""The unique identifier of the data source"""