+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQA457T4YDAY6Y801JKX2QHH"
+        "01KQAEKN32W9S61BGJ1FTJ6MCF"
     )
     r"""The unique identifier of the data source"""