+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQA457T6SDM9EZF46TQVPF2M"
+        "01KQAEKN3AQTF0TG2YXS6DY0J7"
     )
     r"""The unique identifier of the data source"""