+++ packages/orq-rc/RELEASES.md
@@ -7399,3 +7399,13 @@ Based on:
 - [python v4.9.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.7 - packages/orq-rc
+
+## 2026-04-28 16:29:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.8 - packages/orq-rc